feat: add `goleveldb` build flag
because goleveldb won't be the default CometBFT DB in the future.
