+++ goleveldb.go
@@ -1,3 +1,6 @@
+//go:build goleveldb
+// +build goleveldb
+
 package db
 
 import (
